WalletStorage.addPath removed
